adds menu options: Journal > Move Selection Layer Up / Down

+++ src/core/control/layer/LayerController.h
@@ -70,7 +70,7 @@ public:
     void addNewLayer();
     void deleteCurrentLayer();
     void copyCurrentLayer();
-    void moveSelectionToLayer(int layerNo);
+    void moveSelectionToLayer(size_t layerNo);
     void moveCurrentLayer(bool up);
     void mergeCurrentLayerDown();
     void switchToLay(Layer::Index layerId, bool hideShow = false);
